requirements.txt
lightrag-hku[api]
nest-asyncio
pypdf
langchain-unstructured
unstructured
streamlit